tool/pyproject.toml
[project]
name = "tool"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "algoliasearch>=4.16.0",
    "google-api-python-client>=2.166.0",
    "google-api-python-client-stubs>=1.29.0",
    "pillow>=11.2.1",
    "prompt-toolkit>=3.0.50",
    "pydantic>=2.10.6",
    "pydantic-settings>=2.8.1",
    "requests>=2.32.3",
    "ruff>=0.11.2",
    "supabase>=2.15.0",
    "tqdm>=4.67.1",
]

[dependency-groups]
dev = ["ipython>=9.0.2"]
